refactor/#269: SpaceView 에서 BubblePlacer 파일 분리
